refactor back to ftp task
