+++ ramls/movetonode.json
@@ -50,7 +50,7 @@
         }, {
           "$ref" : "#/definitions/FileTask"
         }, {
-          "$ref" : "#/definitions/SftpTask"
+          "$ref" : "#/definitions/FtpTask"
         }, {
           "$ref" : "#/definitions/RequestTask"
         }, {
@@ -206,7 +206,7 @@
             }, {
               "$ref" : "#/definitions/FileTask"
             }, {
-              "$ref" : "#/definitions/SftpTask"
+              "$ref" : "#/definitions/FtpTask"
             }, {
               "$ref" : "#/definitions/RequestTask"
             }, {
@@ -278,7 +278,7 @@
             }, {
               "$ref" : "#/definitions/FileTask"
             }, {
-              "$ref" : "#/definitions/SftpTask"
+              "$ref" : "#/definitions/FtpTask"
             }, {
               "$ref" : "#/definitions/RequestTask"
             }, {
@@ -384,7 +384,7 @@
             }, {
               "$ref" : "#/definitions/FileTask"
             }, {
-              "$ref" : "#/definitions/SftpTask"
+              "$ref" : "#/definitions/FtpTask"
             }, {
               "$ref" : "#/definitions/RequestTask"
             }, {
@@ -709,7 +709,7 @@
       "title" : "FileTask",
       "required" : [ "deserializeAs", "name", "asyncBefore", "asyncAfter" ]
     },
-    "SftpTask" : {
+    "FtpTask" : {
       "type" : "object",
       "additionalProperties" : false,
       "properties" : {
@@ -754,6 +754,9 @@
         "op" : {
           "type" : "string",
           "enum" : [ "GET", "PUT" ]
+        },
+        "scheme" : {
+          "type" : "string"
         },
         "host" : {
           "type" : "string"
@@ -1115,7 +1118,7 @@
             }, {
               "$ref" : "#/definitions/FileTask"
             }, {
-              "$ref" : "#/definitions/SftpTask"
+              "$ref" : "#/definitions/FtpTask"
             }, {
               "$ref" : "#/definitions/RequestTask"
             }, {
@@ -1199,7 +1202,7 @@
             }, {
               "$ref" : "#/definitions/FileTask"
             }, {
-              "$ref" : "#/definitions/SftpTask"
+              "$ref" : "#/definitions/FtpTask"
             }, {
               "$ref" : "#/definitions/RequestTask"
             }, {
