+++ ramls/sftptask.json
@@ -46,6 +46,9 @@
       "type" : "string",
       "enum" : [ "GET", "PUT" ]
     },
+    "scheme" : {
+      "type" : "string"
+    },
     "host" : {
       "type" : "string"
     },
