+++ ramls/subprocess.json
@@ -51,7 +51,7 @@
         }, {
           "$ref" : "#/definitions/FileTask"
         }, {
-          "$ref" : "#/definitions/SftpTask"
+          "$ref" : "#/definitions/FtpTask"
         }, {
           "$ref" : "#/definitions/RequestTask"
         }, {
@@ -219,7 +219,7 @@
             }, {
               "$ref" : "#/definitions/FileTask"
             }, {
-              "$ref" : "#/definitions/SftpTask"
+              "$ref" : "#/definitions/FtpTask"
             }, {
               "$ref" : "#/definitions/RequestTask"
             }, {
@@ -291,7 +291,7 @@
             }, {
               "$ref" : "#/definitions/FileTask"
             }, {
-              "$ref" : "#/definitions/SftpTask"
+              "$ref" : "#/definitions/FtpTask"
             }, {
               "$ref" : "#/definitions/RequestTask"
             }, {
@@ -397,7 +397,7 @@
             }, {
               "$ref" : "#/definitions/FileTask"
             }, {
-              "$ref" : "#/definitions/SftpTask"
+              "$ref" : "#/definitions/FtpTask"
             }, {
               "$ref" : "#/definitions/RequestTask"
             }, {
@@ -722,7 +722,7 @@
       "title" : "FileTask",
       "required" : [ "deserializeAs", "name", "asyncBefore", "asyncAfter" ]
     },
-    "SftpTask" : {
+    "FtpTask" : {
       "type" : "object",
       "additionalProperties" : false,
       "properties" : {
@@ -767,6 +767,9 @@
         "op" : {
           "type" : "string",
           "enum" : [ "GET", "PUT" ]
+        },
+        "scheme" : {
+          "type" : "string"
         },
         "host" : {
           "type" : "string"
@@ -1128,7 +1131,7 @@
             }, {
               "$ref" : "#/definitions/FileTask"
             }, {
-              "$ref" : "#/definitions/SftpTask"
+              "$ref" : "#/definitions/FtpTask"
             }, {
               "$ref" : "#/definitions/RequestTask"
             }, {
@@ -1212,7 +1215,7 @@
             }, {
               "$ref" : "#/definitions/FileTask"
             }, {
-              "$ref" : "#/definitions/SftpTask"
+              "$ref" : "#/definitions/FtpTask"
             }, {
               "$ref" : "#/definitions/RequestTask"
             }, {
